feat: Major indexer refactor with improved reliability and performance
## Core Improvements
- **Complete codebase restructuring**: Split monolithic lagoon_indexer.py into modular components
  - lagoon_event_formatter.py: Dedicated event formatting logic
  - lagoon_event_processor.py: Separated event processing and database operations
  - Improved maintainability and separation of concerns

## Performance & Reliability Enhancements
- **Database transaction atomicity**: Wrapped entire block processing in transactions
  - Prevents partial processing during task cancellation
  - Ensures checkpoint consistency with data integrity
  - Eliminates constraint violations from interrupted operations

- **Optimized event fetching**:
  - Single RPC call for all event types instead of per-type calls
  - Direct topic matching instead of brute-force processing
  - Reduced blockchain API calls by about 90%

- **Robust error handling**:
  - Async retry mechanism with exponential backoff and jitter
  - Graceful cancellation handling in task completion callbacks
  - Automatic provider refresh on failures

## Task Management & Orchestration
- **Dynamic task lifecycle management**:
  - Automatic task creation for new vault deployments
  - Graceful task cancellation for deactivated vaults
  - Prevention of duplicate tasks per vault
  - Continuous operation instead of time-limited runs

- **Multi-vault support**:
  - Concurrent indexing of multiple vaults per chain
  - Independent task tracking and management
  - Database-driven deployment discovery

## Development Experience
- **Auto-restart functionality**:
  - Fixed watchmedo integration in docker-compose.yml
  - Automatic container restart on file changes
  - Proper argument passing with -- separator

## Bug Fixes
- **Event processing fixes**:
  - Fixed AttributeDict assignment errors
  - Corrected enum type mismatch (deposit_request_canceled -> deposit_canceled)
  - Resolved empty logs and duplicate processing issues

- **Database consistency**:
  - Added keeper_bot_enabled field to factory data insertion
  - Fixed return value mismatch in vault snapshot queries
  - Improved error handling in database operations

## Technical Debt
- **Code organization**: Modular architecture for better maintainability
- **Async/await patterns**: Consistent async programming throughout
- **Logging improvements**: Enhanced debugging and monitoring capabilities
- **Configuration**: Environment-driven settings for flexibility

Note: Headroom still needs to be set 2 blocks behind the chain head to avoid missing events
